Replace middleware csp with meta tag csp
